+++ src/Command/CodeCheckerCommand.php
@@ -61,7 +61,7 @@ class CodeCheckerCommand extends AbstractPluginCommand
             $this->standard = $resolver->resolve($this->standard);
         }
 
-        $this->finder = Finder::create()->notPath('yui/build')->name('*.php')->name('*.js')->notName('*-min.js');
+        $this->finder = Finder::create()->notPath('yui/build')->notPath('amd/build')->name('*.php')->name('*.js')->notName('*-min.js');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)